requirements.txt
fastapi==0.111.0
uvicorn
pydantic~=2.7.1
pydantic_settings==2.3.3
python-dotenv==1.0.1
Requests~=2.31.0
starlette==0.37.2
redis~=5.0.6